splits up feature tests to ensure functionality
