+++ test/unit-tests/message-tests.js
@@ -9,14 +9,6 @@ describe('Message', function() {
 
   it('each message should have an id, title, and user', function() {
     var message = new Message({id: 1, user: 'Brett', content: 'Work somewhere with a devops team'});
-
     assert.equal(message.id, 1, message.title, 'Brett Idea', message.content, 'Work on a devops team');
-  });
-
-  it.skip('should update the message text from local storage', function() {
-    var message1 = new Message({ content: 'text1' });
-    var message2 = message1.content('text2');
-    message.saveEditedMessage( message2 );
-    assert.equal(this.message.content, 'text2');
   });
 });
